int literals now is 64bits, rename int literals in AST to 'IntLiteralExpr' + simplify parsing of int literals

+++ stage1/zom_common/src/token.rs
@@ -284,7 +284,7 @@ pub enum TokenType {
     At,        // ` @ `
 
     // Literals
-    Int(i32),
+    Int(u64),
     Float(f32),
     Str(String),
     Char(char),
